Refactor: 1차 구조 변경(모듈화)
