+++ .github/workflows/publish.yaml
@@ -11,4 +11,4 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: subosito/flutter-action@v1
-      - uses: erickzanardo/flutter-gh-pages@v3
+      - uses: erickzanardo/flutter-gh-modules@v3
