+++ lib/global_widgets/image_with_animated_opacity.dart
