+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:easy_dynamic_theme/easy_dynamic_theme.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:portfolio/controller/carousel_index_controller.dart';
-import 'package:portfolio/pages/template_page.dart';
+import 'package:portfolio/modules/template_module/template_page.dart';
 import 'package:portfolio/utils/theme_data.dart';
 import 'package:url_strategy/url_strategy.dart';
