+++ lib/modules/about_module/local_widgets/events_section.dart
