+++ lib/modules/career_module/career_page.dart
@@ -4,7 +4,7 @@ import 'package:portfolio/models/career_model.dart';
 import 'package:portfolio/models/work_model.dart';
 import 'package:portfolio/utils/responsive.dart';
 import 'package:portfolio/utils/url.dart';
-import 'package:portfolio/widgets/app_section.dart';
+import 'package:portfolio/modules/career_module/local_widgets/app_section.dart';
 import 'package:timelines/timelines.dart';
 
 class CareerPage extends StatelessWidget {
