+++ lib/modules/template_module/template_page.dart
@@ -1,15 +1,15 @@
 import 'package:easy_dynamic_theme/easy_dynamic_theme.dart';
 import 'package:flutter/material.dart';
-import 'package:portfolio/pages/about_page.dart';
-import 'package:portfolio/pages/blog_page.dart';
-import 'package:portfolio/pages/career_page.dart';
-import 'package:portfolio/pages/home_page.dart';
-import 'package:portfolio/pages/project_page.dart';
-import 'package:portfolio/pages/skill_page.dart';
+import 'package:portfolio/modules/about_module/about_page.dart';
+import 'package:portfolio/modules/blog_module/blog_page.dart';
+import 'package:portfolio/modules/career_module/career_page.dart';
+import 'package:portfolio/modules/home_module/home_page.dart';
+import 'package:portfolio/modules/project_module/project_page.dart';
+import 'package:portfolio/modules/skill_module/skill_page.dart';
 import 'package:portfolio/utils/responsive.dart';
-import 'package:portfolio/widgets/custom_app_bar.dart';
-import 'package:portfolio/widgets/custom_drawer.dart';
-import 'package:portfolio/widgets/footer.dart';
+import 'package:portfolio/modules/template_module/local_widgets/custom_app_bar.dart';
+import 'package:portfolio/modules/template_module/local_widgets/custom_drawer.dart';
+import 'package:portfolio/modules/template_module/local_widgets/footer.dart';
 import 'package:scrollable_positioned_list/scrollable_positioned_list.dart';
 
 class TemplatePage extends StatelessWidget {
